+++ changelog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [1.1.2] - 2023-06-01
+
+### Added
+ - Added tests for the clients routes 
+ - Added some validation scenarios for the routes
+
+### Fixed
+ - Fixed the status of the create route
+
+
 ## [1.1.0] - 2023-05-31
 
 ### Added